sm5803: disable sinking before other reconfiguration
Silicon Mitus say that writing to flow2 enable bits when flow1 is set to an
active state (sinking or sourcing) will have no effect, and a silently
failed write like that may lead to unexpected (but not dangerous) behavior
because the enable bits will not be in the expected state. Most code paths
already satisfy this, but disabling the sink currently does not: change
sm5803_vbus_sink_enable to disable sinking before updating other
configuration, the mirror of ordering when enabling sinking.

BUG=none
TEST=nereid battery still charges/discharges as expected when PD source is
     connected or disconnected
BRANCH=none
LOW_COVERAGE_REASON=no emulators currently exist for sm5803

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I50c7bf20569d7cad058c19aa238aeed8eaf2aa56