test: host_command: Add coverage to locate_chip
Create a new binary that does not enable the cbi eeprom. This gets
the host command path to return ec_res_unavailable. For the
alternative subcommand, it checks if the TCPC is on the same chip as
the EC (which is not supported in Zephyr) so we can ignore that branch
since it's only for legacy builds and cannot be tested reasonably.

BRANCH=none
BUG=none
TEST=./twister -T zephyr/test/drivers

Signed-off-by: Yuval Peress <peress@google.com>
Change-Id: I1380b882c3fadcd9057a0ed340feb8787bafd80a