flash: Move printing multiple regions to separate function
Printing information about sectors with different size requires
flash_bank_array to be defined. When using Zephyr the array is not
available. In this case we should use structures defined in the
Zephyr flash driver.

Moving code that depends on flash_bank_array will make easier to
to enable support for variable sector sizes in CrosEC with Zephyr.

BUG=b:239712345
BRANCH=none
TEST=make buildall -j

Signed-off-by: Patryk Duda <pdk@semihalf.com>
Change-Id: I5b93a910b9c1c5906da1648a61f00f42e563bf25