twister_tags.py: Fix err if deleted yamls
When a testcase.yaml file gets deleted, twister_tags.py is invoked with
the --validate-files option with no argument. This causes a crash due to
a missing argument argparse exception. Also when twister_tags.py is
invoked with moved testcase.yaml files, it attempts to parse yamls that
no longer exist.

Make --validate-files with no arguments a noop and check if a
testcase.yaml exists before attempting to validate its tags.

BRANCH=none
BUG=none
TEST=repo upload
TEST=./util/twister_tags.py --validate-files

Signed-off-by: Aaron Massey <aaronmassey@google.com>
Change-Id: Idc4f9c44717826d59e59c6ca0627374c1c7eb0e5