[WIP] system: Introduce RW Safe Mode
Introduce a new boot mode called RW Safe Mode. RW Safe Mode runs
after a panic in the RW image.

Capturing a coredump of EC memory when a panic occurs is constrained
by limited EC memory and differing RO/RW memory layouts. RW Safe mode
attempts to work around these limitiations. Essentially, after a
panic in the RW image, the EC resets back to the RW image with
non-critical features disabled in order to preserve as much memory as
possible. Next, the AP is rebooted to depthcharge, which extracts the
EC coredump and saves it to SPI flash. Finally, the EC does a full
reset and boots normally.

BUG=b:249128225
BRANCH=None
TEST=Boot skyrim, crash in RW, observe RW Safe Mode starting

Change-Id: Ida5560504fcdf044c96ecd669867f30c050bb445
Signed-off-by: Rob Barnes <robbarnes@google.com>