motion_sense: fix data race of collection_rate
The motion_sensor.collection_rate record the next sensor
sampling timeout. collection_rate are mostly touched in
motino_sense_task, but when suspend/resume or shutdown,
the variable maybe touched by the hooks.  This will cause the
race condition when the calibration_rate in if-condition is checked in
motion_sensor_time_to_read and then entering the if-clause modified,
and the calibration_rate is now set to 0 by the hooks, this will trigger
the DIVIDE-BY-ZERO exception in increment_sensor_collection.

This CL fix the data race by locking the calibration_rate before
referencing calibration_rate.

BUG=b:251069579
TEST=suspend/resume for 1000+ times and no EC crashes.
BRANCH=kukui

Change-Id: I05e67fe07e63610d612157aeabc73b618bc3bf1e
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>