zephyr: shim: eSPI: enable host interface interrupts explictly
In PR:51013, we drop the enabling of host interface interrupts during
driver initialization because it makes more sense to enable interrupts
after we set the eSPI configuration and register callbacks.
This CL implements the logic above in the eSPI shim by calling eSPI
API to enable interrupts.

BRANCH=none
BUG=b:246963730
TEST=in npcx9_evb, test the kbc/pm channel/port80 communication are
normal between EC and host emulator.
TEST=in skyrim, the hostcmd broken after sysjump symptom is fixed by
the related PR and CL.

Signed-off-by: Jun Lin <CHLin56@nuvoton.com>
Change-Id: Ieea4fd6f52b809110c835df227443d368cf12093