pwm: Factor out PWM_RAW/PERCENT_TO_PERCENT/RAW()
The shim and common pwm code both define the exact same macros to
convert to PWM raw and percentage values.

Factor these macros out as a single macro included in the pwm.h header
that is already included by both the shim and common code. This has the
other side-effect of exposing these macros for testing purposes.

BRANCH=none
BUG=b:236132469
BUG=b:236132199
TEST=zmake build -a
TEST=make runhosttests
TEST=CQ

Signed-off-by: Aaron Massey <aaronmassey@google.com>
Change-Id: I8a8c19242b0122a1a18c26bba3eda08a0613d824