uart: chargen: Fix chargen cmd on Zephyr
The chargen command is hanging the console because it
attempts to stop the thread that it is running in. The shell
thread should only be stopped by a separate thread (it should not
attempt to stop itself).

Modify chargen to read from the shell's input buffer instead of
attempting to stop the shell thread.

BUG=b:251872865
BRANCH=none
TEST=chargen 36 50000 (run to completion)
     chargen 36 50000 (interrupt by sending 'x')
     Verify EFS2 boot is still functional
     Tested on Skyrim(Zephyr) and Nipperkin(legacy)

Change-Id: If536a53f646846f307a7a49dfde993329950b0c5
Signed-off-by: Mark Hasemeyer <markhas@google.com>