tcpmv2: dpm: add message response wait state
This CL adds the MSG_WAIT_RESP state to the DPM state machine. This
state is entered when the DPM needs to send either a VDM for mode
entry/exit or a USB4 enter mode message.

This state removes the need for an explicit sync bit with the PE
layer, and so DPM_FLAG_PE_SYNC is removed as well. When the PE
completes sending the message and waiting for the reply, the DPM is
notified and this triggers an exit back to the appropriate ready
state.

BUG=b:194504052
BRANCH=none
TEST=Verified that mode entry/exit is successful for DP, TBT, and
USB4 on Voxel.
ectool typeccontrol <port> 0 -> exit mode
ectool typeccontrol <port> 2 0 -> DP
ectool typeccontrol <port> 2 1 -> TBT
ectool typeccontrol <port> 2 2 -> USB4

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I09e8faf5e8394f8ab0f961e1df613072d9fa4139