ec: New tool to merge lcov files
Added new tool to merge lcov files using one file as the template and
including only lines that are present in the template file.

The name is because you use the template file like a "stencil" and the
matching coverage shows through. Not great, but it's not really a
"set intersect" or a "merge" either.

Switch firmware_builder.py and gitlab to use the new tool in place of
the filename filtering.

See
https://jbettis.users.x20web.corp.google.com/www/herobrine_rpt/index.html
for updated coverage output.

BRANCH=None
BUG=None
TEST=Ran firmware_builder.py locally.

Signed-off-by: Jeremy Bettis <jbettis@google.com>
Change-Id: I0e9672c346971b0df4602b0adce27fea6367c6b4