test: EC_CMD_HOST_SLEEP_EVENT bad arg
Add a test that verifies the EC_CMD_HOST_SLEEP_EVENT host command may be
invoked with an invalid sleep event and not return an error to the
caller. Also verify that this invocation results in the bad event being
propagated down to chip specific event handler with zero sleep
transitions.

BRANCH=none
BUG=b:252887178
TEST=twister -s zephyr/test/drivers/drivers.power_host_sleep

Signed-off-by: Aaron Massey <aaronmassey@google.com>
Change-Id: I0f177dc660746740b5380254583df426c9b903ca