uart: chargen: Fix chargen cmd on Zephyr
The chargen command is hanging the console because it
attempts to stop the thread that it is running in. The shell
thread should only be stopped by a separate thread (it should not
attempt to stop itself).

Modify chargen to execute as a deferred task in the work queue.  Stop
the console(legacy) or activate the console bypass feature(Zephyr) while
chargen is running to prevent it from hijacking reads of the 'x'
character.

Add support for the Zephyr EC console in uart_stress_tester.py

BUG=b:251872865
BRANCH=none
TEST=chargen 36 50000 (run to completion)
     chargen 36 50000 (interrupt by sending 'x')
     uart_stress_tester.py -t 10 -d <raw_ec_uart>
     Tested on Skyrim(Zephyr)
     Tested on Nipperkin(legacy with and without this commit in fw)

Change-Id: If536a53f646846f307a7a49dfde993329950b0c5
Signed-off-by: Mark Hasemeyer <markhas@google.com>