util: add script useful to parse thousands of crash reports
This CL adds a script that is useful to parse thousands of crash
reports.

It is useful to parse watchdog crash reports where only the PC is
provided. And by analyzing thousands of watchdog crash reports, chances
are that we find the function that triggers the watchdogs.

This CL should be used together with this other script:
https://source.corp.google.com/piper///depot/google3/experimental/users/ricardoq/crashpad/main.py
which lives in Google3 since it has to perform an SQL query.

BUG=None
TEST=crash_analyzer.py lite -m ~/tmp/rammus_193.map -f /tmp/dumps
     And it generated valid report.
BRANCH=None

Change-Id: I954c8b89a01dd2b55c07e19111b3ba887b119215
Signed-off-by: ricardoq@chromium.org