hooks: Update rc handling for deferred calls
The old Zephyr workqueue api was deprecated in 2.6, and recenty removed:
https://github.com/zephyrproject-rtos/zephyr/pull/46812
The hooks module is using the new api, but the return code handling
needs updating.

k_work_schedule() replaces k_delayed_work_submit_to_queue(), and its
return values are different.
In the old api, `-EINVAL` is returned when a "Work item is being
processed or has completed its work.":
https://docs.zephyrproject.org/2.5.0/reference/kernel/threads/index.html#c.k_delayed_work_submit_to_queue

In the new api, a value >= 0 is returned when work has been (or is)
properly scheduled:
https://docs.zephyrproject.org/3.2.0/kernel/services/threads/workqueue.html#c.k_work_schedule_for_queue

Update the return code handling in hook_call_deferred() to correctly
process the return values produced by the new workqueue api.

BUG=b:251872865
BRANCH=none
TEST=`chargen 36 50000` with chromium:3943262 cherry-picked

Change-Id: I4511e9bd5a9f50caa3473e78619ab042026970ca
Signed-off-by: Mark Hasemeyer <markhas@google.com>