gpio: Optionally get the set output instead of the input
Add a GPIO configuration bit to indicate whether a GPIO's input
is unavailable, and instead return the value set as an output.

The input may be unavailable for a number of reasons, such as
the GPIO being an output only GPIO, or used as an open drain with
a low pull-up voltage.

The default for GPIOs should be to always read the value of the pin,
except when this flag is set.

BUG=b:254057906
TEST=./twister -T zephyr/test/drivers
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: I63353586279c755334dd17d8119bfc26b751426c