corsola: set HDMI HPD in interrupt handler
HPD pulse has to toggle with 0.5ms ~ 1ms. Previously,
this was toggled in the hook, which will be hard to
compliant to the requirement. This CL moves the HPD
set function from hook to the interrupt handler to
reduce the delay.

Also, fixed the bugs when HDMI is plugged on boot and
sometimes won't be detected. This is due to the usb_mux_set
function will be skipped before we intentionally makes the
USB_PD_PORT_COUNT to 2 after task inited. We fix this by
re-configure mux setting after tasks inited.

BUG=b:252999628
TEST=1. it6505 is able to capture the HPD interrupt when unplugged
     2. HDMI port is able to output when plugged on boot
BRANCH=none

Change-Id: I6347033e2a7996a8f3a972e9f6b3feb328d4f8c7
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>