gpio: npcx: fix incorrect report from gpio_get_flags_by_mask
The current driver reports the configured level regardless of the GPIO's
direction (input/output). It should report it only when the pin in
question is an output. Also, it should read the level from the PxDOUT
register instead of the PXDIN register.

BRANCH=none
BUG=b:254739925
TEST=build npcx9_evb with CONFIG_GPIO_GET_EXTENDED and
CONFIG_CMD_GPIO_EXTENDED defined, check the report form `gpioget`
console command is correct.

Signed-off-by: Jun Lin <CHLin56@nuvoton.com>
Change-Id: I1f40bb757c1a012697374b294b57934608d5e293