TCPMv2: Add host command to send VDM REQ messages
Add an option for the AP to send us VDM REQ messages to pass on to the
port partner or cable.  This is the first of a series of commits, and
later ones will add the transmission of replies to the AP as well as the
underlying EC actions that still need to take place (ex. storing active
modes and such).

BRANCH=None
BUG=b:208884535
TEST=./twister -T ./zephyr/test

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I205eaa73e7b4b7aa64c1168ef1d8505f1781a1cd