chargen: test: Add chargen unit testing
The chargen console command does not currently have any test coverage.

Add unit testing for chargen. Disable the command's UART_INTERRUPT_DRIVEN
dependency as it is unnecessary, and the test uart is not interrupt
driven. Also add ring buffer support to a couple tests that need it now
that chargen is enabled.

BUG=b:251872865
BRANCH=none
TEST=./twister -v -i -T ./zephyr/test/drivers
     ./twister -v -i --coverage -p native_posix -p unit_testing

Change-Id: I7bff3caf751ddb51b5c73dfec72938725786737e
Signed-off-by: Mark Hasemeyer <markhas@google.com>