npcx: flash: FLASH_BANK_SIZE shouldn't depend on FLASH_ERASE_SIZE
The flash bank defination is used to check the protection block. There
is no reason to define it following the flash erase size setting.
This CL fixed the flash bank size to 64KB which is the max flash erase
block in the SPI flash.

BRANCH=none
BUG=b:254877499
TEST="BOARD=npcx9_evb make", use console command flashwp/flashinfo to
check the flash protection.

Signed-off-by: Jun Lin <CHLin56@nuvoton.com>
Change-Id: I9dfdc4083e2aa0a198ff882ecaa7794bb6ac5ad5