npcx: flash: FLASH_BANK_SIZE shouldn't depend on FLASH_ERASE_SIZE
The flash bank definition is used to check the protection block. There
is no reason to define it following the flash erase size setting.
This CL sets the smallest flash bank size to 64KB, which is 1/8 of 512KB
or 1/16 of 1MB flash.

BRANCH=none
BUG=b:254877499
TEST="BOARD=npcx9_evb make", use console command flashwp/flashinfo to
check the flash protection.

Signed-off-by: Jun Lin <CHLin56@nuvoton.com>
Change-Id: I9dfdc4083e2aa0a198ff882ecaa7794bb6ac5ad5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3980704
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Tested-by: CH Lin <chlin56@nuvoton.com>
Commit-Queue: CH Lin <chlin56@nuvoton.com>