ec: Add #line directive to config_chip.h
There is a problem, where sometimes the coverage for 1 board (so far
always brya RO) will be off by one line. It seems to be related to the
file build-??/zephyr/include/generated/autoconf.h. Adding a #line
directive to the config_chip.h file, which includes autoconf.h, appears
to fix the problem.

BRANCH=None
BUG=b:256008706,b:253131248
TEST=zmake build --coverage brya

Signed-off-by: Jeremy Bettis <jbettis@google.com>
Change-Id: Ib8cc8fca38b5b914cc1f6d87cc74956728327ee6