zephyr: fix incorrect adc channel value in krabby test
Fix an incorrect usage of `adc_emul_const_value_set()`:
The `temp_sensors[CHARGER_TEMP].idx` actually returns the enum-name of
the given channel, not the channel number (i.e. the number specified in
the io-channels prop).

Both numbers are zero in our DTS, so we didn't catch the bug.

BUG=none
TEST=twister -T zephyr/test/krabby
BRANCH=none

Signed-off-by: Ting Shen <phoenixshen@google.com>
Change-Id: Ib786889be72927743c7d4cbe3185514896e3af2e