test: hc_cmd_pd_chip_info: Fix stack smashing
In run_pd_chip_info() test helper function, the BUILD_HOST_COMMAND macro
is passed a pointer to an ec_response_pd_chip_info_v1 struct but expects
the direct value.

Dereference the pointer when in the helper as it is passed into the
BUILD_HOST_COMMAND macro.

BRANCH=none
BUG=none
TEST=twister -i -s zephyr/test/drivers/drivers.host_cmd
TEST=CQ

Signed-off-by: Aaron Massey <aaronmassey@google.com>
Change-Id: I732f6bd53e2ed6bf17dba063eee914ec2525b61d