tpcmv2: dpm: Rename attempt_mode_entry/exit functions
This CL renames these two functions as part of DPM refactoring. In
addition, the functions are moved in the file to be adjacent to the
state handler functions.

This CL does not change any functionality. Instead this change is
intended to help track the changes to these 2 functions for the
subsequent CL which adds a msg_wait state.

BUG=b:194504052
BRANCH=none
TEST=Verified that mode entry/exit is successful for DP, TBT, and
USB4 on Voxel.
ectool typeccontrol <port> 0 -> exit mode
ectool typeccontrol <port> 2 0 -> DP
ectool typeccontrol <port> 2 1 -> TBT
ectool typeccontrol <port> 2 2 -> USB4

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I9ea39368ae819ade972eebb771f863d3c7e5bd58