zephyr: test: Test `crash` console command
Test various `crash` subcommands that cause different types of fatal
errors. Not all were possible to test due to limitations with running
these on a PC. Also had to make a tweak to the busy loops in the
watchdog and hang subcommands so that they don't interfere with the
simulated sys clock in native_posix. Add a public wrapper function for
calling the crash command outside of the shell interface so we don't
mess up its internal state when the command doesn't return. Also make
the crash command compile without optimizations so it stays broken
exactly as intended ;-).

BUG=None
BRANCH=None
TEST=./twister

Signed-off-by: Tristan Honscheid <honscheid@google.com>
Change-Id: Iac1fb5871d78b9c269568f2fdc88148717c15882