common/uart_printf: Change uart_put/uart_put_raw return value
Instead of returning EC_SUCCESS/EC_ERROR_OVERFLOW, return the number of
characters written. This makes retrying for the failures possible and is
compatible with the semantics of write().

This should have no effect on existing code, since uart_put() is not
used anywhere. The only place that uart_put_raw is used, the return
value is ignored.

BRANCH=none
BUG=b:234181908, b:254530679
TEST=make buildall

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: If9e8f01663f160b44bbcc4bb313327486dd0d4d0