tcpmv2: Clear dpm request for data reset based on PD revision
At present dpm request is cleared only for Partner's SOP
PD revision < 3.0. For cables with PD rev 2.0, there is no
response for data reset request, which results into a timeout
causing Protocol Error. This CL is to check for PD revision of
SOP prime along with SOP to decide on the dpm request for
data reset.

BUG=b:257102175
BRANCH=None
TEST=Verified with Samsung 2k TBT display whose cable is PD rev 3.0

Signed-off-by: poornima tom <poornima.tom@intel.com>
Change-Id: I36948c88000ef5dd031012eecbacf04c2f5f832d