board/{ambassador,endeavour,fizz,genesis,moonbuggy}: Adjust fan tables for CfM use-case.
Chromebox for Meetings (CfM) devices are videoconferencing appliances.
Discovering that a significant fraction of our fleet was throttling the
CPU due to temperature, resulting in bad video quality, we have decided
that our devices should run at low-enough temperatures so that they
never enter the throttling condition. The present change reduces the
temperature thresholds at which we increase fan speeds.  For most of the
time, when the CfM is idle, we expect no or little change in fan speeds.
During calls, the CfM CPU can exceed 53 C, and we move all the fan
thresholds to between 54 C and 70 C in an attempt to arrest any
temperature increase before reaching throttling temperature (typically
74 C).

I left the fan speeds the same, in case those have to do with the
controller for each hardware model, and only changed the thresholds.

Comparing with other devices' fan tables, most seem to have transitions
in the 50-60 C range.  By comparison, most of the endeavour and fizz
tables changed by this seemed to allow the fan to stay at the slower
idle rate until the device reached near-throttling temperatures

Additional details at go/cfm-overheating

BRANCH=update-fan-tables
BUG=b:252966838
TEST=Built firmware, flashed to endeavour, verified that fan turns on
earlier

Change-Id: I4eb532977677fbfd939fc46f851948cd4adf9788
Signed-off-by: Richard Yeh <rcy@google.com>