u2f: Fix some constant usage
The size of auth_time_secret_hash should be SHA256_DIGEST_SIZE, not
U2F_AUTH_TIME_SECRET_SIZE, though they are the same now. This will make
the userland check more consistent with the actual constants used in the
header.

Also, the CORP_SALT_SIZE is defined incorrectly and never used, fix it
to the correct value and use it in the struct field size.

BUG=None
TEST=make buildall -j

Change-Id: I22dcf2cfd46819475ca916e0b0c3dcd5b3369b19
Signed-off-by: Howard Yang <hcyang@google.com>