gsctool: fix misleading error message
The getopt_long() library function returns '?' (decimal 63) in case
there is a command line parameter with required value, but there is no
value in the command line.

gsctool is printing "could not find long opt table index for 63" in
this case, which is misleading. This patch changes error message to
make sense.

BUG=none
TEST=before this change running './gsctool  -D -R' results in
   "could not find long opt table index for 63"

   after this change error message is
   "Command line error, parameter argument missing"

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: I5b1a8fda62a7edd673ea1181efb73108677b77d1