Add CONFIG_CHARGER_MAX_INPUT_CURRENT_LIMIT
This splits the dual use of CONFIG_CHARGER_INPUT_CURRENT into two
different symbols, changing the uses of that which are used to set a
minimum current limit to be CONFIG_CHARGER_MAX_INPUT_CURRENT_LIMIT.

Most boards implement this in the same way within either the board or
baseboard, so handling of the new option is moved into
charge_set_input_current_limit (which is called by every user of the
option) and every board which repeated this pattern has the new symbol
set to the same value as the old one, with the duplicated code deleted.

To verify that all changes are appropriate, the following has been
manually checked:
 * All boards with a change to a .c file also have a .h change
 * All boards without a changed .h file have a changed baseboard.h

This is intended to verify that each board that duplicated the MAX()
logic has its configuration updated to use the shared copy, and that
boards with that code in the baseboard also update their configuration.

Change-Id: Ia460a16293c1fb82aac3784fd9be57ba0985f2fe