zephyr: shim: use device specific 1.8V GPIO flags
The generic GPIO_VOLTAGE_1P8 flag is deprecated and has been removed
upstream, in favor of device specific flags, which right now have been
defined to use the same bit, but that may change in the future.

In the EC firmware we share cunks of dts files on boards with different
architecture, so we still need a common flag that can be redefined to
the platform specific one.

This patch supports the change by:
- redefine GPIO_VOLTAGE_1P8 for the devicetree files, but set it for the
  specific architecture being built, this is done using the platform dts
  files
- redefine GPIO_VOLTAGE_1P8 for the source code depending on what GPIO
  controller is used (which would not work if we had to use it on port
  expanders, but that may never be a problem)
- define an entry for native_posix for the unit test to use
- undefine the legacy GPIO_SEL_1P8V entry if not supported

BRANCH=none
BUG=none
TEST=cq dry run

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Cq-Depend: chromium:4023593
Change-Id: Ife34ecc23fb1e18b13a7bcdbe378b0a182c50249