board/fizz, baseboard/kalista: Adjust fan tables for CfM use-case.
Chromebox for Meetings (CfM) devices are videoconferencing appliances.
Discovering that a significant fraction of our fleet was throttling the
CPU due to temperature, resulting in bad video quality, we have decided
that our devices should run at low-enough temperatures so that they
never enter the throttling condition. The present change reduces the
temperature thresholds at which we increase fan speeds.  For most of the
time, when the CfM is idle, we expect no or little change in fan speeds.
During calls, the CfM CPU can exceed 53 C, and we move all the fan
thresholds to between 54 C and 70 C in an attempt to arrest any
temperature increase before reaching throttling temperature (typically
74 C).

Additional details at go/cfm-overheating

Since these boards are shared with non-CfM versions, we shall not apply
this change to main line ChromeOS; but only CP it to every ChromeOS
milestone's CfM branch.

BRANCH=update-fan-tables
BUG=b:252966838
TEST=Confirmed validity of settings with `ectool thermalset`. Built
firmware, flashed to endeavour, verified that fan turns on earlier

Change-Id: I34788cff1663c4b23aed0c2daba7ed47fc41e391
Signed-off-by: Richard Yeh <rcy@google.com>