board/fizz, baseboard/kalista: Adjust fan tables for CfM use-case.
Chromebox for Meetings (CfM) devices are videoconferencing appliances.
Discovering that a significant fraction of our fleet was throttling the
CPU due to temperature, resulting in bad video quality, we have decided
that these devices should never throttle. The present change reduces the
temperature thresholds at which we increase fan speeds.  For most of the
time, when the CfM is idle, we expect no or little change in fan speeds.

During calls, the CfM CPU can exceed 60 C, and the sensors controlling
the fan are typically 10-20 degrees C cooler than the CPU, so we set
the maximum-fan temperature to 50 C in an attempt to arrest any
temperature increase before the CPU reaches throttling temperatures.

Additional details at go/cfm-overheating

Since these boards are shared with non-CfM versions, we shall not apply
this change to main line ChromeOS; but only CP it to every ChromeOS
milestone's CfM branch.

BRANCH=update-fan-tables
BUG=b:252966838,b:191187610,chromium:1383859
TEST=Confirmed validity of settings with `ectool thermalset`. Built
firmware, flashed to fizz, verified that fan turns on earlier
This does not completely prevent throttling from 47.25 W to 15.00 W
on Kaby Lake boards (endeavour, fizz, kalista), but it does prevent
dropping below 15.00 W.

Change-Id: I34788cff1663c4b23aed0c2daba7ed47fc41e391
Signed-off-by: Richard Yeh <rcy@google.com>