fan: Refactor and test the most common custom fan_percent_to_rpm.
This is in preparation for reusing this implementation for
{ambassador, genesis, moonbuggy, scout}.

The below-named boards are all the boards that have defined
CONFIG_FAN_RPM_CUSTOM.

{kalista, berknip, chronicler, dewatt, endeavour, ezkinil, fizz} have
similar implementations of a path-dependent fan_percent_to_rpm
function. The path-dependence enables hysteresis, typically to avoid
oscillations in temperature and fan speed.  (fizz includes excelsior)

{chronicler} performs boxcar smoothing of the input.

{osiris} has two fans, and other boards have multiple sensors, and we
have left their refactoring for another time.

BRANCH=refactor-fan-percent-to-rpm
BUG=b:252966838,b:191187610,chromium:1383859
TEST=make -j buildall && make -j runhosttests

Signed-off-by: Richard Yeh <rcy@google.com>
Change-Id: I50ad4d78ac1145f92573a417646c1f57b8945463