test: test memory mapped storage config
Add a new binary variant for memory mapped storage. Add tests for
All the edge cases around the flash memory as well as verify that
reads and writes work. As done in other tests, an override is added
to common/flash.c allowing us to provide our own buffer to intercept
the flash address.

BRANCH=none
BUG=none
TEST=twister

Signed-off-by: Yuval Peress <peress@google.com>
Change-Id: Icca8ac74a6c98a93b995162efb87b3f5fa111a8a