npcx9: fix SHA256 ROM API table address
It was found that the hardware SHA256 accelerator is broken. This is
caused by the table address of SHA ROM API changed from 0x100 to 0x13C
since the ES2 chip.
This CL fixed the issues described above.

BRANCH=none
BUG=b:155771688
TEST=burn the same RW image; #define/#undef CONFIG_SHA256_HW_ACCELERATE;
verify the RW hash value is the same in the console message.
TEST=with the test CL:2455022, move test patterns in test/sha256.c
to board/npcx9_evb/test_sha256.c; pass all test patterns.

Signed-off-by: Jun Lin <CHLin56@nuvoton.com>
Change-Id: I12ce8531a925f02474f66d90bf7efd950b3eeba3