zephyr: leds: move led value in the gpio led binding
The current cros-ec,gpio-led-pins binding is leaking values into the
named-gpios binding. That's an incorrect way of using the devicetree and
only works because named-gpios is not currently using the value cell and
because we control the binding.

Move the values in a separate property within the cros-ec,gpio-led-pins
binding.

BRANCH=none
BUG=b:242826402
TEST=validated the resulting structure with gdb
TEST=zmake build -a; ./twister

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: I77a09b25cc69c5faed254da38316104b304c8f5e