zephyr: herobrine: Pass through AUX termination when DP-Config is acked
The hardware block diagram is like:

          AUX termination
                 |
            +----+--- TCPC MUX --- PPC SW --- Type-C connector (port-0)
DP PHY -- DP MUX
            +----+--- TCPC MUX --- PPC SW --- Type-C connector (port-1)
                 |
          AUX termination

The current implementation muxes the path end-to-end on a HPD high
event.

The dependencies are:
 * Chromebook muxing the AUX channel depends on a HPD high event,
 * DP sink raising HPD depends on a cable connection detected (auto-
   probe),
 * Successful probe of a cable connection depends on AUX termination,
which cause a dead-lock.

This change breaks the dead-lock. During the DP-Config phase, when the
DP-Config VDM is acked, connect the PPC SW and TCPC MUX, but leave the
DP MUX disconnected. So the DP sink can detect the correct AUX
termination. The DP port selection mux is not enabled yet so no signal
passes through to the DP PHY.

As the TCPC MUX is enabled in the DP-Config phase, we can't use the TCPC
MUX status as the condition to check if the DP path is muxable. Change
it to the DP selection mux.

BRANCH=None
BUG=b:247005613
LOW_COVERAGE_REASON=No existing test for the board-specific code
TEST=Together with the next CL.
Connect the device through a Type-C dongle to a monitor and verify
the screen. Try the combinations:
 * Plug C0 monitor -> plug C0 dongle -> screen on C0 monitor
 * Plug C0 dongle -> plug C0 monitor -> screen on C0 monitor
   -> unplug C0 monitor -> plug C0 monitor -> screen on C0 monitor
 * Plug C0 dongle -> plug C0 monitor -> screen on C0 monitor
   -> plug C1 dongle -> -> plug C1 monitor -> no screen on C1 monitor
   -> unplug C0 monitor -> replug C1 monitor -> screen on C0 monitor
 * Plug C0 dongle -> plug C1 dongle -> plug C1 monitor
   -> screen on C1 monitor -> unplug C1 dongle -> plug C0 monitor
   -> screen on C0 monitor

Change-Id: I5a5e08dc26ebaa22ff7a3cfc0828ed2e01df3995
Signed-off-by: Wai-Hong Tam <waihong@google.com>